修改音量显示，以及更改低分辨率的被拉伸显示

+++ mediaPlayer/app/src/main/java/com/gq/mediaplayer/MyPlayer.java
@@ -376,7 +376,7 @@ class MyPlayer implements MediaPlayer.OnBufferingUpdateListener,
                 }
             } else {
                 float scaleVideo = videoWidth / (float) videoHeight;
-                float scaleSurface = width / height;
+                float scaleSurface = width / (float) height;
                 if (scaleVideo > scaleSurface) {
                     params.width = width;
                     params.height = (int) (width / scaleVideo);
